new: to basic types, address mapping
