+++ src/YaoBlocks.jl
@@ -15,6 +15,7 @@ include("composite/composite.jl")
 include("algebra.jl")
 include("blocktools.jl")
 include("layout.jl")
+include("treeutils/treeutils.jl")
 
 include("deprecations.jl")
